refactor: 메서드 시그니처 변경

+++ src/main/java/racingcar/dao/CarDao.java
@@ -2,5 +2,5 @@ package racingcar.dao;
 
 public interface CarDao {
 
-    void insert(final String name, final int position, final int gameId, final boolean isWin);
+    void insert(final String name, final int position, final Long gameId, final boolean isWin);
 }
